api: add pagination token to QueryMeta
API queries can request pagination using the `LastToken` and `PerPage`
fields of `QueryOptions`, when supported by the underlying API.

Add a `LastToken` field to the `structs.QueryMeta` so that we have a
common field across RPCs to tell the caller where to resume paging
from on their next API call. Include this field on the `api.QueryMeta`
as well so that it's available for future versions of List HTTP APIs
that wrap the response with `QueryMeta` rather than returning a simple
list of structs.

Mark `NextToken` as deprecated so that no new callers use it; the sole
caller of this are CSI external API calls and we can move the field
into those specific request structs.